Add a very nasty solution for having ETA's
Also, patch the tests. The nastiness here is mostly caused by our
terrible state handling and stats gathering / passing along. I know it
leaves a lot of refactoring pending with some code duplication and long
expressions. But I'm happy that we now have an skeleton on which to
work on. After all, this already works.